move code out of API.Client: it's needed without a client
