+++ API.Client/DlcsClient.cs
@@ -7,6 +7,7 @@ using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 using DLCS.HydraModel;
+using DLCS.Web.Auth;
 using DLCS.Web.Response;
 using Hydra;
 using Hydra.Collections;
