+++ Orchestrator/Infrastructure/Deliverator/DeliveratorApiClient.cs
@@ -7,6 +7,7 @@ using API.Client;
 using DLCS.Core.Encryption;
 using DLCS.Core.Types;
 using DLCS.Model.Customers;
+using DLCS.Web.Auth;
 using DLCS.Web.Requests;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
