+++ Portal/Behaviours/AuditBehaviour.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using API.Client;
 using DLCS.Mediatr.Behaviours;
+using DLCS.Web.Auth;
 using MediatR;
 using Microsoft.Extensions.Logging;
