+++ Portal/Features/Admin/Requests/DeleteSignupLink.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using API.Client;
 using DLCS.Model.Customers;
 using DLCS.Repository;
+using DLCS.Web.Auth;
 using MediatR;
 using Microsoft.Extensions.Logging;
