+++ Portal/Features/Admin/Requests/GetAllSignupLinks.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using API.Client;
 using DLCS.Repository;
+using DLCS.Web.Auth;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
