+++ Portal/Features/Spaces/Requests/ToggleManifestMode.cs
@@ -5,6 +5,7 @@ using API.Client;
 using DLCS.Model.Spaces;
 using DLCS.Repository;
 using DLCS.Repository.Entities;
+using DLCS.Web.Auth;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Portal.Behaviours;
