+++ Portal/Pages/Shared/_Layout.cshtml
@@ -1,4 +1,5 @@
 ﻿@using API.Client
+@using DLCS.Web.Auth
 <!doctype html>
 <html lang="en">
   <head>
