+++ Portal/Pages/Spaces/Details.cshtml.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using API.Client;
 using DLCS.Core.Settings;
 using DLCS.HydraModel;
+using DLCS.Web.Auth;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
